Refactoring, moving database functions to various.py
